--- a/src/mpi/pt2pt/testall.c
+++ b/src/mpi/pt2pt/testall.c
@@ -44,12 +44,10 @@ int MPIR_Testall_impl(int count, MPIR_Request *request_ptrs[], int *flag,
     mpi_errno = MPID_Progress_test();
     if (mpi_errno != MPI_SUCCESS) goto fn_fail;
 
-    for (i = 0; i < count; i++)
-    {
+    for (i = 0; i < count; i++) {
         if (request_ptrs[i] != NULL &&
                 request_ptrs[i]->kind == MPIR_REQUEST_KIND__GREQUEST &&
-                request_ptrs[i]->u.ureq.greq_fns->poll_fn != NULL)
-        {
+                request_ptrs[i]->u.ureq.greq_fns->poll_fn != NULL) {
             mpi_errno = (request_ptrs[i]->u.ureq.greq_fns->poll_fn)(request_ptrs[i]->u.ureq.greq_fns->grequest_extra_state,
                     &(array_of_statuses[i]));
             if (mpi_errno != MPI_SUCCESS) goto fn_fail;
@@ -109,8 +107,8 @@ program to unexecpectedly terminate or produce incorrect results.
 .N MPI_ERR_REQUEST
 .N MPI_ERR_ARG
 @*/
-int MPI_Testall(int count, MPI_Request array_of_requests[], int *flag, 
-		MPI_Status array_of_statuses[])
+int MPI_Testall(int count, MPI_Request array_of_requests[], int *flag,
+                MPI_Status array_of_statuses[])
 {
     MPIR_Request * request_ptr_array[MPIR_REQUEST_PTR_ARRAY_SIZE];
     MPIR_Request ** request_ptrs = request_ptr_array;
@@ -124,7 +122,7 @@ int MPI_Testall(int count, MPI_Request array_of_requests[], int *flag,
     MPIR_FUNC_TERSE_STATE_DECL(MPID_STATE_MPI_TESTALL);
 
     MPIR_ERRTEST_INITIALIZED_ORDIE();
-    
+
     MPID_THREAD_CS_ENTER(GLOBAL, MPIR_THREAD_GLOBAL_ALLFUNC_MUTEX);
     MPIR_FUNC_TERSE_PT2PT_ENTER(MPID_STATE_MPI_TESTALL);
 
@@ -133,21 +131,21 @@ int MPI_Testall(int count, MPI_Request array_of_requests[], int *flag,
     {
         MPID_BEGIN_ERROR_CHECKS;
         {
-        int i = 0;
+            int i = 0;
 
-	    MPIR_ERRTEST_COUNT(count, mpi_errno);
+            MPIR_ERRTEST_COUNT(count, mpi_errno);
 
-	    if (count != 0) {
-		MPIR_ERRTEST_ARGNULL(array_of_requests, "array_of_requests", mpi_errno);
-		/* NOTE: MPI_STATUSES_IGNORE != NULL */
-		MPIR_ERRTEST_ARGNULL(array_of_statuses, "array_of_statuses", mpi_errno);
-	    }
-	    MPIR_ERRTEST_ARGNULL(flag, "flag", mpi_errno);
+            if (count != 0) {
+                MPIR_ERRTEST_ARGNULL(array_of_requests, "array_of_requests", mpi_errno);
+                /* NOTE: MPI_STATUSES_IGNORE != NULL */
+                MPIR_ERRTEST_ARGNULL(array_of_statuses, "array_of_statuses", mpi_errno);
+            }
+            MPIR_ERRTEST_ARGNULL(flag, "flag", mpi_errno);
 
-	    for (i = 0; i < count; i++) {
-		MPIR_ERRTEST_ARRAYREQUEST_OR_NULL(array_of_requests[i], i, mpi_errno);
-	    }
-	}
+            for (i = 0; i < count; i++) {
+                MPIR_ERRTEST_ARRAYREQUEST_OR_NULL(array_of_requests[i], i, mpi_errno);
+            }
+        }
         MPID_END_ERROR_CHECKS;
     }
 #   endif /* HAVE_ERROR_CHECKING */
@@ -156,32 +154,28 @@ int MPI_Testall(int count, MPI_Request array_of_requests[], int *flag,
     MPIR_CHKLMEM_DECL(1);
 
     /* Convert MPI request handles to a request object pointers */
-    if (count > MPIR_REQUEST_PTR_ARRAY_SIZE)
-    {
+    if (count > MPIR_REQUEST_PTR_ARRAY_SIZE) {
         MPIR_CHKLMEM_MALLOC_ORJUMP(request_ptrs, MPIR_Request **,
                 count * sizeof(MPIR_Request *), mpi_errno, "request pointers", MPL_MEM_OBJECT);
     }
 
     n_completed = 0;
-    for (i = 0; i < count; i++)
-    {
-        if (array_of_requests[i] != MPI_REQUEST_NULL)
-        {
+    for (i = 0; i < count; i++) {
+        if (array_of_requests[i] != MPI_REQUEST_NULL) {
             MPIR_Request_get_ptr(array_of_requests[i], request_ptrs[i]);
             /* Validate object pointers if error checking is enabled */
 #           ifdef HAVE_ERROR_CHECKING
             {
                 MPID_BEGIN_ERROR_CHECKS;
                 {
-                    MPIR_Request_valid_ptr( request_ptrs[i], mpi_errno );
+                    MPIR_Request_valid_ptr(request_ptrs[i], mpi_errno);
                     if (mpi_errno) goto fn_fail;
                 }
                 MPID_END_ERROR_CHECKS;
             }
 #           endif
         }
-        else
-        {
+        else {
             request_ptrs[i] = NULL;
             n_completed += 1;
         }
@@ -190,24 +184,19 @@ int MPI_Testall(int count, MPI_Request array_of_requests[], int *flag,
     mpi_errno = MPID_Testall(count, request_ptrs, flag, array_of_statuses);
     if (mpi_errno != MPI_SUCCESS) goto fn_fail;
 
-    for (i = 0; i < count; i++)
-    {
-        if (request_ptrs[i] != NULL)
-        {
-            if (MPIR_Request_is_complete(request_ptrs[i]))
-            {
+    for (i = 0; i < count; i++) {
+        if (request_ptrs[i] != NULL) {
+            if (MPIR_Request_is_complete(request_ptrs[i])) {
                 n_completed++;
                 rc = MPIR_Request_get_error(request_ptrs[i]);
-                if (rc != MPI_SUCCESS)
-                {
+                if (rc != MPI_SUCCESS) {
                     if (MPIX_ERR_PROC_FAILED == MPIR_ERR_GET_CLASS(rc) || MPIX_ERR_PROC_FAILED_PENDING == MPIR_ERR_GET_CLASS(rc))
                         proc_failure = TRUE;
                     mpi_errno = MPI_ERR_IN_STATUS;
                 }
             } else if (unlikely(MPIR_CVAR_ENABLE_FT &&
                         MPID_Request_is_anysource(request_ptrs[i]) &&
-                        !MPID_Comm_AS_enabled(request_ptrs[i]->comm)))
-            {
+                        !MPID_Comm_AS_enabled(request_ptrs[i]->comm))) {
                 mpi_errno = MPI_ERR_IN_STATUS;
                 MPIR_ERR_SET(rc, MPIX_ERR_PROC_FAILED_PENDING, "**failure_pending");
                 status_ptr = (array_of_statuses != MPI_STATUSES_IGNORE) ? &array_of_statuses[i] : MPI_STATUS_IGNORE;
@@ -217,51 +206,34 @@ int MPI_Testall(int count, MPI_Request array_of_requests[], int *flag,
         }
     }
 
-    if (n_completed == count || mpi_errno == MPI_ERR_IN_STATUS)
-    {
+    if (n_completed == count || mpi_errno == MPI_ERR_IN_STATUS) {
         n_completed = 0;
-        for (i = 0; i < count; i++)
-        {
-            if (request_ptrs[i] != NULL)
-            {
-                if (MPIR_Request_is_complete(request_ptrs[i]))
-                {
+        for (i = 0; i < count; i++) {
+            if (request_ptrs[i] != NULL) {
+                if (MPIR_Request_is_complete(request_ptrs[i])) {
                     n_completed ++;
                     status_ptr = (array_of_statuses != MPI_STATUSES_IGNORE) ? &array_of_statuses[i] : MPI_STATUS_IGNORE;
                     rc = MPIR_Request_complete(&array_of_requests[i], request_ptrs[i], status_ptr, &active_flag);
-                    if (mpi_errno == MPI_ERR_IN_STATUS && status_ptr != MPI_STATUS_IGNORE)
-                    {
+                    if (mpi_errno == MPI_ERR_IN_STATUS && status_ptr != MPI_STATUS_IGNORE) {
                         if (active_flag)
-                        {
                             status_ptr->MPI_ERROR = rc;
-                        }
                         else
-                        {
                             status_ptr->MPI_ERROR = MPI_SUCCESS;
-                        }
                     }
-                }
-                else
-                {
-                    if (mpi_errno == MPI_ERR_IN_STATUS && array_of_statuses != MPI_STATUSES_IGNORE)
-                    {
+                } else {
+                    if (mpi_errno == MPI_ERR_IN_STATUS && array_of_statuses != MPI_STATUSES_IGNORE) {
                         if (!proc_failure)
                             array_of_statuses[i].MPI_ERROR = MPI_ERR_PENDING;
                         else
                             array_of_statuses[i].MPI_ERROR = MPIX_ERR_PROC_FAILED_PENDING;
                     }
                 }
-            }
-            else
-            {
+            } else {
                 n_completed ++;
-                if (array_of_statuses != MPI_STATUSES_IGNORE)
-                {
+                if (array_of_statuses != MPI_STATUSES_IGNORE) {
                     MPIR_Status_set_empty(&array_of_statuses[i]);
                     if (mpi_errno == MPI_ERR_IN_STATUS)
-                    {
                         array_of_statuses[i].MPI_ERROR = MPI_SUCCESS;
-                    }
                 }
             }
         }
@@ -270,13 +242,12 @@ int MPI_Testall(int count, MPI_Request array_of_requests[], int *flag,
     *flag = (n_completed == count) ? TRUE : FALSE;
 
     /* ... end of body of routine ... */
-    
+
   fn_exit:
-    if (count > MPIR_REQUEST_PTR_ARRAY_SIZE)
-    {
+    if (count > MPIR_REQUEST_PTR_ARRAY_SIZE) {
         MPIR_CHKLMEM_FREEALL();
     }
-    
+
     MPIR_FUNC_TERSE_PT2PT_EXIT(MPID_STATE_MPI_TESTALL);
     MPID_THREAD_CS_EXIT(GLOBAL, MPIR_THREAD_GLOBAL_ALLFUNC_MUTEX);
     return mpi_errno;
@@ -285,9 +256,9 @@ int MPI_Testall(int count, MPI_Request array_of_requests[], int *flag,
     /* --BEGIN ERROR HANDLING-- */
 #   ifdef HAVE_ERROR_CHECKING
     {
-	mpi_errno = MPIR_Err_create_code(
-	    mpi_errno, MPIR_ERR_RECOVERABLE, FCNAME, __LINE__, MPI_ERR_OTHER, "**mpi_testall",
-	    "**mpi_testall %d %p %p %p", count, array_of_requests, flag, array_of_statuses);
+        mpi_errno = MPIR_Err_create_code(
+            mpi_errno, MPIR_ERR_RECOVERABLE, FCNAME, __LINE__, MPI_ERR_OTHER, "**mpi_testall",
+            "**mpi_testall %d %p %p %p", count, array_of_requests, flag, array_of_statuses);
     }
 #   endif
     mpi_errno = MPIR_Err_return_comm(NULL, FCNAME, mpi_errno);