# Introduction

This is a proposal to refactor `MPI_Wait` family (Wait/Waitall/Waitany/Waitsome) and `MPI_Test` family (Test/Testall/Testany/Testsome) implementations so an MPICH device can override them. By  doing so, a device may for example implement more fine-grain wait on a particular request by polling a limited set of hardware context.

The key idea is that `MPI_Wait*/Test*` call `MPID_Wait*/Test*`, which a device defines. If the device doesn't want to bother implementing its own, it may simply call `MPIR_Wait*_impl/Test*_impl` as a fallback.

Following a standard convention, `MPID_Wait*/Test*` take `MPIR_Request *` **object pointers** instead of `MPI_Request` **handles**. This requires refactoring of existing `MPIR_Wait*_impl/Test*_impl` as well, as they currently take handles.

Replaces #2708 

# Change Summary

This PR essentially includes the following changes:
* Refactors `MPIR_Wait*_impl/Test*_impl`.
    * Change argument type from request handle to request object pointer
    * They are now responsible only for making a progress and checks for completion
    * They used to translate handles, free the request object after completion, nullify the handle, and set the user status. Now these are caller's responsibility except for a very limited situation (e.g. generalized request.)
* Modifies `MPI_Wait*/Test*` so they call `MPID_Wait*/Test*`
* Introduce `#define` macro definitions in CH3/CH4 that redirect `MPID_Wait*/Test*` to `MPIR_Wait*_impl/Test*_impl`

# Potential Drawbacks

Especially for calls that handle multiple requests at once (e.g. `MPI_Waitall`), the `impl` functions typically used to process requests in a single loop, in which everything including handle translation, progress, object deallocation etc. happened. With this PR we had to split this into multiple loops in different functions, there may be performance overhead. Hope is that more optimized device-specific implementation could offset it somehow.

Here are some results from osu_mbr_mr message rate with CH4 on KNL machines (Intel(R) Xeon Phi(TM) CPU 7210 @ 1.30GHz) with OPA-1. osu_mbr_mr uses `MPI_Waitall` to wait for `Isend/Irecv` to complete. Since `MPI_Waitall` has an optimized path when `MPI_STATUES_IGNORE` is used, I also modified the benchmark to specify `MPI_STATUES_IGNORE` (the original code wasn't using it.) Results for 8-byte messages are reported, and these numbers are averages of 10 runs.

## Results with original osu_mbr_mr (no `MPI_STATUSES_IGNORE`)
```
window size  64    8192
------------------------
master   916535  773122
this PR  910998  739364
```

## Results with modified osu_mbr_mr (with `MPI_STATUSES_IGNORE`)
```
window size  64    8192
-----------------------
master   940524  802396
this PR	 935359  775312
```

## Detailed Configurations

* Compiler: ICC 17.0.4
* MPICH configuration
```
  $ /home/hajimefu/src/mpich-review/configure -C --prefix=/home/hajimefu/opt/mpich-psm2-yes-tg-intel --disable-perftest --with-libfabric=/home/hajimefu/opt/libfabric-psm2 --disable-ft-tests --with-fwrapname=mpigf --with-filesystem=ufs+nfs --enable-timer-type=linux86_cycle --enable-romio --with-mpe=no --with-smpcoll=yes --with-assert-level=0 --enable-shared --enable-static --enable-error-messages=yes --enable-visibility --enable-large-tests --enable-g=none --enable-timing=none --enable-error-checking=no --enable-fast=all,O3 --disable-debuginfo --with-device=ch4:ofi:psm2 --enable-handle-allocation=default --enable-threads=multiple --without-valgrind --enable-timing=none --enable-ch4-shm=yes --enable-thread-cs=global --with-ch4-netmod-ofi-args= MPICHLIB_CFLAGS=-O3 -Wall -ggdb  -msse2 -msse4.2 -mcrc32 -mavx512f -ggdb -Wall -mtune=generic -std=gnu99 MPICHLIB_CXXFLAGS=-O3 -Wall -ggdb  -msse2 -msse4.2 -mcrc32 -mavx512f -ggdb -Wall -mtune=generic MPICHLIB_FCFLAGS=-O3 -Wall -ggdb  -msse2 -msse4.2 -mcrc32 -mavx512f -ggdb -Wall -mtune=generic MPICHLIB_F77FLAGS=-O3 -Wall -ggdb  -msse2 -msse4.2 -mcrc32 -mavx512f -ggdb -Wall -mtune=generic MPICHLIB_LDFLAGS=-O3 -L/usr/lib64  -msse2 -msse4.2 -mcrc32 -mavx512f -mtune=generic
```

Another potential drawback is that I haven't tested the ULFM feature which were scattered throughout `MPIR_Wait*_impl/Test*_impl`.