@hajimefu 2017-09-12
Fixed a build failure in CH3.