@hajimefu 2017-09-13
@raffenet in my local testing I didn't see any regression in CH4/OFI and CH3/Nemesis/TCP. Could you invoke a test as well to confirm?