@raffenet 2017-09-14
test:jenkins/ch3/all
test:jenkins/ch4/all