@hajimefu 2017-10-02
Note for myself: incorporate #2761 into this PR as well.