@halimamer 2017-11-21
it's hard to follow this PR because of the large number of whitespace changes that obfuscate the true changes. @hajimefu , could you remove the whitespace changes so that I take a look again?