@hajimefu 2017-11-21
Just updated the branch to
1. rebase onto the latest master
2. incorporate changes happened in #2761

@halimamer Whitespace changes are only in "Fix coding style of MPI_Test*/Wait* functions". Does it help if you ignore this last patch?