@halimamer 2017-11-21
yeah, I can follow better the changes of individual commits