@hajimefu 2017-11-29
Discussed this with @halimamer and his concern was the performance degradation. Added a patch to inline post-processing of `MPI_Waitall` (`mpi: inline waitall post-processing `). Same technique could be applied to other functions too but for now I'm just looking at the performance of it.

```
Without MPI_STATUES_IGNORE
window size  64    8192
------------------------
master	900839	751073
this PR	913105	731721

With MPI_STATUES_IGNORE
window size  64    8192
------------------------
master	927490	782224
this PR	937883	758606
```
Interestingly, now this refactoring beats the master in 64 windows (=64 requests to wait at once), but the master still wins in 8192. This trend seems consistent across ICC (18.0.1)/GCC (6.3.1), KNL/Skylake-SP, and LTO/non-LTO builds. (Haven't tried ICC LTO (IPO)). Numbers here are from ICC, non-LTO builds. Configurations are the same as above, except that now ICC is 18.0.1 and baseline is the master as of this week.

How do we proceed? Do we keep looking at this performance difference for 8192 windows? @halimamer @raffenet