@pavanbalaji 2017-11-30
The important question here is *why* one is better than the other.  Aren't both schemes essentially equivalent when the device does not do any special override?