@hajimefu 2017-11-30
They are equivalent in a logic but implementations are slightly different. Currently `MPIR_Waitall_impl` essentially looks like this:
```C
MPIR_Wait_impl(MPI_Request array_of_requests[])
{
    /* 1. handle translation */
    foreach (array_of_requests) {
        array_of_request_ptrs[i] = get_ptr(array_of_requests[i]);
    }

    foreach (array_of_request_ptrs) {
        /* 2. making progress */
        while (!complete(array_of_request_ptrs[i]))
            MPID_Progress_wait();
        /* 3. post prosessing */
        array_of_statues[i].status = array_of_request_ptrs[i].status;
        MPL_free(array_of_request_ptrs[i]);
        array_of_requests[i] = MPI_REQUEST_NULL;
    }
}

MPI_Waitall(MPI_Request array_of_requests[])
{
    MPIR_Waitall_impl(array_of_requests);
}
```
In this refactored code, `MPIR_Wait_impl (=MPID_Wait)` is only responsible for "2. making progress" part. So the entire code looks like this:
```C
MPIR_Waitall_impl(MPIR_Request array_of_request_ptrs[])
{
    foreach (array_of_request_ptrs) {
        /* 2. making progress */
        while (!complete(array_of_request_ptrs[i]))
            MPID_Progress_wait();
    }
}

MPI_Waitall(MPI_Request array_of_requests[])
{
    /* 1. handle translation */
    foreach (array_of_requests) {
        array_of_request_ptrs[i] = get_ptr(array_of_requests[i]);
    }

    /* 2. making progress */
    MPIR_Waitall_impl(array_of_request_ptrs);

     /* 3. post prosessing */
    foreach (array_of_requests) {
        array_of_statues[i].status = array_of_request_ptrs[i].status;
        MPL_free(array_of_request_ptrs[i]);
        array_of_requests[i] = MPI_REQUEST_NULL;
    }
}
```
So the current code has two loops over an array of requests, whereas the refactored one has three.