@pavanbalaji 2017-11-30
@hajimefu we'll need more analysis than that.  Are using three loops instead of two causing more cache misses or some other issue that's creating worse performance?  There's something more detailed that's happening and without a deep analysis to identify the problem, we won't know what to fix.