@danghvu 2018-01-24
Here are some PAPI counters: instructions count (`TOT_INS`) and L1 data cache miss (`L1_DCM`) per `MPI_Waitall` that I have profiled for this benchmark in our KNL cluster. Ran with windows = 8192, and did not modify to ignore status:

```
master branch
size | MB/s | Messages/s | TOT_INS | L1_DCM
1 | 0.78 | 779323.83 | 1163436.04 | 2781.15
2 | 1.56 | 780292.07 | 2326846.47 | 5629.61
4 | 3.15 | 786695.38 | 3490256.54 | 8484.65
8 | 6.34 | 791960.4  | 4653666.75 | 11321.55

this pull request
size | MB/s | Messages/s | TOT_INS | L1_DCM
1 | 0.77 | 772205.74 | 1204033.79 | 5069.38
2 | 1.53 | 766035.08 | 2408041.65 | 10116
4 | 2.98 | 746067.09 | 3612049.53 | 15103.13
8 | 6.1  | 761932.77 | 4816057.53 | 20157.45
```

L1 misses for this pull request is ~2x the master branch.
So right now my guess is that the original code has better temporal locality i.e. when a request is completed, if it is freed/post-processed immediately (like the original) then you can save a miss since that request is still in cache.