@hajimefu 2018-01-26
@danghvu Thank you for taking a look at this. Decoupling request completion and post processing was one of the decisions we made to clearly separate responsibilities for MPI layer and MPID layer. But apparently that design comes up a non-negligible overhead.

How does it look like then the window size is 64?