@danghvu 2018-01-26
@hajimefu The performance impact and profiling differences are small when there are lesser requests -- the variations are high too (up to 4096 there is almost no difference -- more sample numbers below). 

I have a PR blocked depending on this PR. I'm not sure if it's worth to try something else or just to ignore this (because realistically who would MPI_Waitall on 8192 requests?). Anyway, some idea:
 - processing the list of requests in batch of 64 or less ?
 - move more responsibility to the MPID_ layer e.g .. MPIR_Request_complete/request_complete_fastpath/setting the status .. to lessen the impact ?

```
window = 64
<master>
Size | MB/s | Messages/s | TOT_INS | L1_DCM
1 | 0.87 | 871996.45 | 9723.78 | 19.19
2 | 1.78 | 889660.88 | 19420.28 | 37.24
4 | 3.75 | 937226.57 | 29116.80 | 55.60
8 | 7.31 | 913461.12 | 38813.30 | 73.86

<pr>
Size | MB/s | Messages/s | TOT_INS | L1_DCM
1 | 0.89 | 886386.69 | 9940.47 | 23.20
2 | 1.81 | 904826.96 | 19853.69 | 46.15
4 | 3.67 | 917355.87 | 29766.89 | 67.70
8 | 7.30 | 911918.77 | 39680.12 | 88.24

window = 2048
<master>
Size | MB/s | Messages/s | TOT_INS | L1_DCM
-- | -- | -- | -- | --
1 | 0.86 | 857426.33 | 297268.29 | 676.58
2 | 1.75 | 876395.84 | 594510.29 | 1356.71
4 | 3.55 | 887787.76 | 891752.26 | 2034.07
8 | 7.02 | 877305.98 | 1188994.27 | 2702.47

<pr>
Size | MB/s | Messages/s | TOT_INS | L1_DCM
1 | 0.87 | 865697.80 | 291598.10 | 694.26
2 | 1.73 | 866040.14 | 583169.71 | 1387.30
4 | 3.48 | 870861.11 | 874741.23 | 2081.12
8 | 6.99 | 874190.00 | 1166312.68 | 2771.62

window = 4096
<master>
Size | MB/s | Messages/s | TOT_INS | L1_DCM
1 | 0.83 | 832321.17 | 582197.69 | 1451.05
2 | 1.63 | 815680.82 | 1164369.77 | 2884.98
4 | 3.31 | 827535.80 | 1746542.11 | 4331.54
8 | 6.83 | 853421.68 | 2328714.18 | 5791.01
<pr>
Size | MB/s | Messages/s | TOT_INS | L1_DCM
1 | 0.83 | 831275.38 | 593499.86 | 1952.38
2 | 1.67 | 836159.50 | 1186974.44 | 3907.44
4 | 3.3 | 825739.75 | 1780449.68 | 5880.12
8 | 6.55 | 818752.92 | 2373924.36 | 7837.51


```