@pavanbalaji 2018-01-27
@danghvu we shouldn't ignore the performance difference, even though you might be eager to get to the other PR.  If cache misses is indeed the reason (which it seems to be based on your analysis), then batching requests, like you suggested, seems like the right solution.  Can you fixup this PR to do that?